clean up, restructuring
